+++ API/DTOs/PowerPlantOutput.cs
@@ -3,6 +3,6 @@ namespace API.DTOs
     public class PowerPlantOutput
     {
         public string Name { get; set; }
-        public double P { get; set; }  
+        public decimal P { get; set; }  
     }
 }
